Make a cross platform proto project
