+++ cli/src/main/scala/org/bykn/bosatsu/PlatformModule.scala
@@ -1,9 +1,19 @@
 package org.bykn.bosatsu
 
+import _root_.bosatsu.{TypedAst => proto}
 import cats.MonadError
 import cats.effect.IO
 import com.monovore.decline.Argument
 import java.nio.file.{Path => JPath}
+import java.io.{
+  FileInputStream,
+  FileOutputStream,
+  BufferedInputStream,
+  BufferedOutputStream
+}
+import scalapb.{GeneratedMessage, GeneratedMessageCompanion}
+
+import cats.syntax.all._
 
 object IOPlatformIO extends PlatformIO[IO, JPath] {
   type F[A] = IO[A]
@@ -33,11 +43,64 @@ object IOPlatformIO extends PlatformIO[IO, JPath] {
       }
     )
 
-  def readPackages(paths: List[Path]): IO[List[Package.Typed[Unit]]] =
-    ProtoConverter.readPackages(paths)
+  def read[A <: GeneratedMessage](
+      path: Path
+  )(implicit gmc: GeneratedMessageCompanion[A]): IO[A] =
+    IO.blocking {
+      val f = path.toFile
+      val ios = new BufferedInputStream(new FileInputStream(f))
+      try gmc.parseFrom(ios)
+      finally {
+        ios.close
+      }
+    }
+
+  def write(a: GeneratedMessage, path: Path): IO[Unit] =
+    IO.blocking {
+      val f = path.toFile
+      val os = new BufferedOutputStream(new FileOutputStream(f))
+      try a.writeTo(os)
+      finally {
+        os.close
+      }
+    }
+
+  def readInterfacesAndPackages(
+      ifacePaths: List[Path],
+      packagePaths: List[Path]
+  ): IO[(List[Package.Interface], List[Package.Typed[Unit]])] =
+    (
+      ifacePaths.traverse(read[proto.Interfaces](_)),
+      packagePaths.traverse(read[proto.Packages](_))
+    ).tupled
+      .flatMap { case (ifs, packs) =>
+        IO.fromTry(
+          ProtoConverter.packagesFromProto(
+            ifs.flatMap(_.interfaces),
+            packs.flatMap(_.packages)
+          )
+        )
+      }
 
   def readInterfaces(paths: List[Path]): IO[List[Package.Interface]] =
-    ProtoConverter.readInterfaces(paths)
+    readInterfacesAndPackages(paths, Nil).map(_._1)
+
+  def readPackages(paths: List[Path]): IO[List[Package.Typed[Unit]]] =
+    readInterfacesAndPackages(Nil, paths).map(_._2)
+
+  def writeInterfaces(
+      interfaces: List[Package.Interface],
+      path: Path
+  ): IO[Unit] =
+    IO.fromTry(ProtoConverter.interfacesToProto(interfaces))
+      .flatMap(write(_, path))
+
+  def writePackages[A](packages: List[Package.Typed[A]], path: Path): IO[Unit] =
+    IO.fromTry {
+      packages
+        .traverse(ProtoConverter.packageToProto(_))
+        .map(proto.Packages(_))
+    }.flatMap(write(_, path))
 
   def unfoldDir: Option[Path => IO[Option[IO[List[Path]]]]] = Some {
     (path: Path) =>
